map page on index
